requirements.txt
torch>=2.0
transformers>=4.40
peft>=0.10
datasets>=2.18
pyyaml>=6.0
tqdm>=4.60
accelerate>=0.28
